resources/currency/controller/app/requirements.txt
pylint==2.4.0
google-cloud==0.34.0
google-cloud-secret-manager==0.2.0
google-cloud-storage==1.19.1
Flask==1.1.2
kafka-python==2.0.1

alpha-vantage==2.1.3
gunicorn==20.0.4
cloudevents==0.2.4
